Added auth to remove page.

+++ blogript/remove/remove.js
@@ -6,11 +6,22 @@ form.addEventListener("submit", async (event) =>
 
 	const urlParams = new URL(window.location.toLocaleString()).searchParams;
 
+	if (sessionStorage.userid === undefined)
+	{
+		window.location.replace("/login");
+		return;
+	}
+
 	try
 	{
+		let data = {};
+		data.user_id = sessionStorage.userid;
+
 		response = await fetch("/api/blog/" + urlParams.get("postid"),
 		{
-			method: "DELETE"
+			method: "DELETE",
+			headers: { "Content-Type": "application/json" },
+			body: JSON.stringify(data)
 		});
 		if (!response.ok) throw response;
 		window.location.replace("/");
